PP-6010 Improve how we match on arrays in pact files
Previously, if the response body in a pact contained an array, we would
only match on the type for each element in the array. This meant that
for arrays of object, we would only verify that each element was an
object, rather than matching on each field in the object.

Modify logic for generating pact files so that for arrays of objects, we
add matchers for the fields in the object.

This means that, for example the following:

```
"matchingRules": {
  "$.body.events[0]": {
    "match": "type"
  },
  "$.body.events[1]": {
    "match": "type"
  }
}
```

becomes:

```
"matchingRules": {
  "$.body.events[0].amount": {
    "match": "type"
  },
  "$.body.events[0].state.status": {
    "match": "type"
  },
  ...
  "$.body.events[1].amount": {
    "match": "type"
  },
  "$.body.events[1].state.status": {
    "match": "type"
  },
  ...
}
```