Make pact a dev dependency
In 2f5e2e06 we switched from the old version of pact to the new version,
but it looks to me like we accidentally (?) made it a production
dependency, when it should be a dev dependency (?).

This needs reviewing by someone who understands pact a bit more than I do so
they can confirm my suspicion that it should only be a dev dependency.