PP-5956 Display whether a payment is MOTO on transaction details page
If the `allow_moto` flag is true for the gateway account, show a 'MOTO:'
row on the transaction details page with a value of 'Yes' or 'No'.

If the `allow_moto` flag is false for the gateway account, do not show
the 'MOTO:' row on the transaction details page.

Add a Cypress test for this behaviour.