PP-6156 Add metadata to page for individual payment link
On the page for each individual payment link, add a section for
reporting columns (metadata associated with the payment link),
complete with links to add a new reporting column and edit the
existing ones. Reporting columns are sorted alphabetically
(case-insensitively) by metadata key.

The reporting columns section is only shown if the
experimentalFeaturesEnabled flag is set on the service.

While the reporting columns section uses the summary list component
from the GOV.UK Design System, it uses HTML rather than the
Nunjucks macro. This is because the number of rows in the table is
dynamic (it’s produced by iterating over the metadata keys) and
there’s no nice way to tell the summary list component Nunjucks
macro that we want a variable number of rows (we’d have to pull
the construction of the rows object out of the template itself,
which is complicated).