PP-6156 Check for reporting columns on edit payment link page
On the edit payment link page, check for the existence of the
reporting columns and the button to add new ones. This test is
carefully designed to check the columns are presented in
alphabetical order case-insensitively and that the URLs to change
each piece of metadata correctly URL-encode the key name.