PP-6466 Fix google pay settings page
At some point, we stopped sending the merchant ID to connector. This
means that although we were changing the `allow_google_pay` setting on
the gateway account, it never appeared as enabled because when connector
returns the setting, it always checks that the merchant ID is there.

Also fix an issue where it was not possible to turn off google pay due
to the `required: "true"` attribute on the merchant ID field. Instead
do server side validation for this.