BAU Fix showing flash container when there is a side navigation
Due to the way the grid formatting of the main content and the flash
container were done separately, when the flash container is taller than
the side navigation, the main content overflows into a new row and is
left aligned. This looks very bad.

Fix this by creating three new layouts which extend layout.njk
These are:
layout-with-side-nav.njk
layout-two-thirds-no-nav.njk
layout-full-width.njk

In these layouts, place the flash container and the main content block
in the same grid column. Also move the grid column control for the side
navigation into the layout.

Then for all templates, extend the relevant layout.