PP-6487: payout service for grouping payouts by date
According to the discovery design document for the payout MVP, we will
be displaying payouts for all services grouped by date.

It is also necessary to provide the services name for each payout entry,
as this information isn't available to the payout (from Ledger) there
will also be a requirement to compliment payouts with the service name.

This adds a service helper method to accept a flat list of payouts and return a
grouped object indexed by date.

If a user object is provided (see User model), this will also add the
service name to each payout object according to the gateway account
assigned to that payout.

Example usage:
```
groupPayoutsByDate([{
  id: '...'
  paid_out_date: '2019-09-01T01...'
})

/*
{
  '2019-09-01': {
    date: Moment(),
    entries: [{
      id: '...'
    }]
  }
}
*/
```