PP-6627: patch calling open Connector route with no accounts
The default behaviour in newer backend routes is require a gateway
account id to be passed in to any resource, passing in nothing `key=[]`
or
in the case of comma separated values `key=` will either filter the data
for no account ids (resulting in nothing returned) OR return `400`
requiring something to be set or non empty values.

In the case of the frontend accounts route being called in this utility,
passing no accounts will treat the route as if there is no filter and
return all accounts.

This behaviour will be patched in the backend as part of this ticket.

Patch this behaviour and avoid an unnecessary call to the backend by
simply returning nothing if no accounts will have valid permissions.