PP-6660: multi serivce permissions move all qs parsing to clients
The mulit service permissions library was previously responsible for
preparing the actual query string that was then passed to the backend
for filtering. This caused an issue where it was using a format unknown
by the standard backend routes.

* move all parsing of query strings to the clients
* all parsing of params is done by the `qs` library according to the
same formats, there should be no manual string interpolation
* multi service permissions library simply reduces to a list of account
ids, this is what should be passed around through the code and accepted
by the clients
* rename methods and values in multi service permissions library to
better reflect what is actually being passed around