PP-6638 Redirect to account details redirect route
If the `ENABLE_ACCOUNT_STATUS_PANEL` environment variable is set to
true, redirect to the new `/stripe/add-psp-account-details` route which
determines the next page in the add account details flow to show the
user.

If the feature flag is not enabled, redirect to the dashboard as was
done previously.

Refactor edited tests to use async/await and reduce the number of lines
of test code a bit.