PP-6634 Update content of add stripe account details pages
- Updated content and field ordering on pages to match new design
- Disabled client side validation on the VAT number page as it didn't
seem to be working, it wasn't displaying the error message. This is
possibly because there is now a hint for the field and the JS doesn't
cope well with that. We already abandoned doing client side validation
for other pages in the flow, and are generally moving away from doing
it.