PP-6672 Add live account requested banner
If the user has requested a live account, and a live account has not
yet been created, display a banner on the dashboard.

Only show the banner if the user has the permission 'go-live-stage:read'
which indicates they are an admin with permission to make a request to
go live.