PP-6644 Set default from date for unfiltered transaction search
- When a user visits the `/transactions` page, the search that is run
automatically only lists transactions from 1 month ago. This will use
the same day in the previous month, or if that date does not exist then
the last day of the previous month.
- When a user clears the "From date" filter and clicks "Filter", the
search is performed with no from date limit.
- The "Clear Filter" button has been renamed to "Reset filters" and will
revert to searching for unfiltered transactions limited to 1 month ago.