PP-6717 Allow editing organisation number for all services
Previously, it was only possible to edit if the service had a direct
debit account.

However, we collect the phone number as part of the go live process as
we send it to Stripe.

Make this editable on the organisation details page. Also make it
mandatory. As we require it for accounts that are going live anyway, it
seems acceptable to always require it.