PP-6716: add integrity checks for CDN scripts
Addressing 'x-domain script includes'; add an integrity check to ensure
the file we receive from the CDN is the one expected.

Future work may see these files removed entirely (in favour of HTML
standards) or bundled and served by the application.

- [ ] verify that the SHA-384 signatures for the file line up with what
is being served by the CDN
- [ ] ensure the transaction search date selection still works on common
browsers and no console errors or warnings are shown