PP-6716: specifiy CORS attribute for SRI
When the browser does an integrity check it will use the cross origin
policy of the script. By default this is `same-site` which assumes the
resource is being served by the server hosted domain.

As this is not the case with the CDN, specify `anonymous` to allow the
integrity check to validate the resource.