PP-6864 Remove gateway account stub with repeat behaviour
The repeat behaviour was required because we were setting up all the
stubs in the `beforeEach` for a `describe` block in a Cypress test.

By move the stub setup to the individual `it` blocks where the stubs
will be called, we can avoid this behaviour and make the stub setup less
confusing.