PP-6834: add government banking option to go live
Allow users to select the government banking Worldpay contract using the
constant added in https://github.com/alphagov/pay-adminusers/pull/836.

At the time of writing, the flow for this option should follow the same
behaviour as the bring your own contract Worldpay selection, the
differntiation will be for the Support roles benefit.