PP-6864 Replace test for responsible person template logic
We had a cypress test that tested an edge case of template logic that
both address line 1 and address line 2 display the same error summary in
the summary list, and if there are errors for both fields, it is only
displayed once.

To do this test, we had to type a very long string for address line 2 as
this is the only error that can be simulated.

As this is template logic, it is suited to a template test. So replace
the cypress test with some simple template tests for this.