PP-7002 Stop passing markup in flash messages
We're using flash messages to display errors for certain journeys.
These are the payment link journey and editing a user's phone number.

To avoid passing markup in, use "error" as the flash key rather than
"genericError" and construct the error summary in each template that
expects an error. We are already using the flash message to display the
inline error message on the form.

This isn't our favoured pattern but it removes the anti-pattern of
controllers constructing the HTML in the quickest way.