PP-7002 Make generic error messages use design system
Previously, error messages displayed using the "genericError" flash
message would just be a red box either containing a paragraph of text,
or inconsistently formatted using markup passed in a string from the
controller.

Now make it use the design system error summary component with the title
"There is a problem".