PP-7002 Display errors better for create payment link journey
Previously we were passing validation errors from the POST controller
when we redirect to the GET controller in flash messages in a confusing
way.
These also contained HTML markup constructed by the controller which is
not advised.

Make this work in a more sensible way by putting the validation errors
as well as the submitted form values into the session cookie in a
"recovered" object. Then in the GET controller after redirect, retrieve
from here and delete the object from the session.

Create a nunjucks macro "errorSummary" which takes a list of errors and
displays them in the design system error summary component when
required. This simplifies the repeated template logic. And can be used
in other existing places where we show an error summary.