PP-7002 Display errors better for edit payment link journey
This was previously improved for the create payment link journey.
Improve for the edit payment link journey too.

In addition, we don't need to get the product for each controller as
we get it on the index page and put it in the session. After that on
each edit page, we will always retrieve the fields from the session data
rather than the retrieved product anyway, so it was pointless.

Also rewrite the tests for the get controllers as proper unit tests, as
they were doing weird things with the Accept header on the request in
order to return data being passing to the template rather than the
rendered page.