PP-7214: refund button design system prevent double click
Add the `preventDoubleClick` attribute to the refund payment button.
This "debounces" event propegation for clicks for 1 second after the
button is pressed.

This can only prevent subsequent clicks for 1 second.

This is a temporary fix which will help solve double clicking however as
this is a terminal POST request a further script should be added to
fully disable the button until the action is completed.