PP-5236: api key create requests pass account type
Include the account type for the gateway account that is requesting a
new API key. This is applied to both account API keys and API keys
associated with payment links for the account.

Entirely internal and ad-hoc features like demo payments to test with
your users are not included -- with this set up API keys generated with
those features will not provide and the token account type and will not
expect a prefix on the resulting token.

Update mocks to guarantee the account type is being correctly propegated
through to public auth.