PP-7409: add test vs. live content for payment link created
Payment links created in test and live environments have very different
use cases; the wording for the payment link creation references the
payment link being "live" which conflicts with our use of this status
for accounts.

Add a separate path for test and live accounts to provide users
additional information if they are just testing.