PP-7416 Add button to enable or disable Worldpay 3DS Flex
Add a button to the ‘Your PSP’ page for Worldpay accounts to enable
or disable 3DS Flex.

The enable button is shown only if it’s a Worldpay account, 3DS is
enabled, 3DS Flex is disabled and there are 3DS Flex credentials.

The disable button is shown only if it’s a Worldpay account, 3DS
is enabled and 3DS Flex is enabled.

No button is shown in other circumstances, meaning, for example,
that a service cannot enable 3DS Flex if they have not entered
3DS Flex credentials.

The page that’s requested when the button is pressed requires
the toggle-3ds:update permission whereas the page it’s displayed
on requires the gateway-credentials:read permission. This is a
less than ideal but only the admin role has either of these so
the button won’t be shown to anyone who can’t press it.

Remove the ‘Remove credentials’ button from the 3DS Flex
credentials page (which was the only way of turning off 3DS Flex
before) because that’s no longer necessary (some code to support
this function is still there; it will be removed in a future
commit).