PP-7458 Order service by ID
At some point the ordering of services on the “My services page”
changed. We weren’t doing any explicit ordering of services so were just
relying on the order returned by the database.

This could have been when we bulk updated the “created_date” recently,
but it's unclear.

Explicitly order on the id as we think that’s how services were
previously ordered.