PP-7360: move payment links metadata handling out to util
Multiple actions across reporting columnts interact with adding,
updating and removing metadata in different ways. Move this out to a
shared util so that it can be tested in one place.

Also add a session interface to abstract consistently getting the
payment links session. In the future this can be based on the request
object itself to know if you are creating or updating a payment link.