PP-7488 Ensure all "My profile" routes run authorisation middleware
Group "/my-profile/" paths in object within paths object.

Previously, "/my-profile/phone-number" wasn't added to the authenticated
paths array, so not all authorisation middleware was included on this
route. This means we weren't checking that the user has a valid session
and we weren't applying CSRF checks to the POST route.

Add `paths.profile` to the authenticated routes array so all
paths under it have the authorisation middleware and CSRF middleware
applied.