PP-7488 Add unit test to ensure authorisation middleware used
Add test to ensure that authorisation middleware is called for all
routes that require authentication.

Achieve this by specifying in the test which paths do not require
authentication. Then check all route setup calls to ensure that the
authorisation middleware is used for routes with paths that are not in
this list.

This will ensure that the middleware is used for any new routes that
are registered, unless they are explicitly added to the exclude list.