PP-7530: fix reporting columns URL encode key regression
In order to try and support the current and proposed implementation of
reporting columns a regression was introduced with no longer URL encoded
the reporting column key when editing. This is then used in a path and
causes unexpected behaviour when escape characters are used.

This fixes the issues with URLs so special characters that were blocked
for that reson have been removed.