PP-7509: always delete backend product metadata when in-flow metdata
When the flag is enabled for reporting columns that are managed in the
session flow, consistently remove the backend product metadata as it
won't include the changes that have been made since the edit session has
started.

The other values that are managed in the flow are "merged" with the
backend product every time the payment link "edit" page is loaded, this
works because these values will always be replaed. As reporting columsn
can be edited and even entirely removed, the session version should be
considered the true copy until submission.