PP-7588 Demonstrate how pact dependency will be removed from fixtures
Want to remove the pact dependency so that fixtures can be used directly
by Cypress test code, which cannot use Node dependencies.

The pact dependency is currently used to construct pacts for responses
which include the appropriate matchers.

To do this:
- Remove `.getPactified()` methods returned by fixtures. Instead,
"pactify" the response in the pact test itself.
- Add utility to return "pactifiers" to pactify certain responses.
- Remove `getPlain()` and `getObject()` methods returned by fixtures
and just return the raw response object. We can construct the models
returned by `getObjecT()` wherever we need them easily enough.