PP-7588 Remove pact dependency from transactions fixtures
Remove dependency on pact-base which depends on the npm module
@pact-foundation/pact/

Use the pactifier utility to pactify any responses in pact test
constructed using the fixtures.

Calls to getPlain() for the fixtures are removed as the fixtures now
return a plain object.

Also remove all unused fixtures from `transaction.fixtures.js` - these
exist from when connector was used to retrieve transactions rather than
ledger. The remaining fixtures that are still used are all for refunds,
so rename to `refund.fixtures.js`.