PP-7588 Move gateway account stubs
Move gateway account stubs out of stubs.js and construct them directly
in the gateway-account-stubs helper file.

Remove some stubs for patch requests which were never actually hit due
to mis-configuration of the stubs. As these stubs just return 200
without a body they are unnecessary as Mountebank returns a 200 by
default for any request that does not match a stub.