PP-7583 Remove direct debit code
Remove all code for getting direct debit accounts. This includes code
for the direct debit account dashboard and linking a direct debit
account to Go Cardless.

We are not currently supporting the direct debit integration and much of
the code for this has already been removed.

In the re-architecting of the middleware for resolving accounts for URLs
we are not supporting the direct debit integration as we do not know
what this integration will look like if we resume work on it.

The only code that remains related to direct debit is checking whether
an account has a direct debit id before proceeding, but this can
probably be removed too.