PP-7583: fix inconsistency in setting session gateway account id
Previously the gateway account id has always been set from the admin
users service role list, this will always be a string as admin users
also stores external ids in the same way.

The proposed structure now uses middleware that fetches the gateway
account directly from Connector, providing a `number` id.

While we are using both methodologies (and moving over to the
proposed) we will need to normalise this value into a String to make
sure the checks by existing middleware continue to succeed.

This fixes an issue with the service being incorrectly reset to the
default service after the existing middleware failed to find the newly
set gateway account id as it was comparing a number to a string.