PP-7583 Move 3ds routes to use account URL structure
Use the account router for these routes so that the URL contains the
account external ID and we use our new middleware for resolving the
service/account and checking the user is authorised.

Also move paths for existing URLs that are now using the account router
to be under `account` in the Object exported by `paths.js`.