PP-7583 Remove unnecessary getEmailNotifications middleware
This middleware made an unnecessary call to connector to get the account
when this has already been retrieved in previously run middleware.

It then duplicated properties already on the account object with new
names.

Ideally we should be using a model to restructure the data in an
appropriate way when we retrieve the account from connector, but for now
just access this account data from the original place it lives on the
get account response in controllers that use it.