PP-7445: add gateway accounts url structure upgrade util
Simple utility to upgrade or redirect account specific URLs from the
current method to the proposed URL structure.

This will both rederict users who are following older links, allow a
seamless release for users actively using the admin tool during a
release and log upgrades so that we can see how often this is happening
and when legacy code can be removed.