PP-7445: set account external id on session when switching accounts
The only valid reason for switching an account with the current service
switcher methodology should be through the my services page.

This ensures the external ID is set by the service switcher, the
external ID is then used when upgrading legacy account URLs to the new
structure.

With the existing middleware that no longer meets best practice, it is
possible to assign a new gateway account by default (the 0th gateway
account belonging to a given user). This is difficult to understand both
from a user and developer perspective and will be removed with the
legacy code when the new structure is in.