PP-7445: make sure account external id is available on login
As we transition URL structure for accounts only new middelware routes
and the service switcher will be setting the external id required to
redirect to the correct URLs.

Ensure the external ID is set when the user logs in, this should then be
correct until the ID is changed by a known redirect controller.