PP-7583 Move settings route
Un-reverts a previously reverted commit to update the settings route to
use the new account URL structure.

It was previously reverted due to the potential for the user to see 404s
while we were deploying because the new route would not exist on the old
instances during deploy, but new instances would redirect to the new
route. The new route now already exists so that won't be an issue.