PP-7583 Fix zap tests
Add back in the api-keys path to the authenticated routes array so that
we redirect to the login page when someone tries to access the old
`/api-keys` URL directly without having a gateway account cookie set.

This is to temporarily fix ZAP tests, we will find a better solution to
this.