PP-7677 Add account dashboard route
Add a route that goes to the account dashboard for a specific account
using the new account URL structure.

Make it so that the index route `/`, which previously called the
dashboard controller now uses the account ID it has retrieved from the
old `getAccount` middleware it runs to redirect to the new dashboard
route.

Update existing redirects so that they redirect to the new dashboard
route if we know what account the user is in.

Continue to redirect to `/` in the following cases. When we default to
the "My services page", this will redirect the user to there:
- After the user has logged in
- After the user has created a service as part of registration

Add a simple Cypress test to test the redirect from `/` to the account
dashboard.

Remove the following functional tests as they are unnecessary:

test/integration/get-gateway-account.ft.test.js
- for old middleware - is covered in other places

test/integration/service-switch.ft.test.js
- is covered by integration tests in service-switch.controller.it.test.js