PP-7667: dashboard activity navigate directly to page
Fix a minor regression in the dashboard activity page. The test coverage
for this feature navigates directly to the page however the template
currently relies on the page being redirected from the root index `/`.

Directly submit from requests to the dashboard page for now to line up
with the test coverage. How this should be covered should be considered
following up.