PP-7561 Show logged in header on error pages
If the user is logged in, display the logged in header containing
"My services", "My profile" etc links. If they are logged out display
the logged out header.

Achieve this by making the error pages extend `layout.njk` and rendering
using the display converter.

In the display converter, set the `loggedIn` variable which is used by
`header.njk` and `footer.njk` based on whether the user is logged in.
Previously this was always set to true in the `layout.njk` base template
and the error pages extended `layout-logged-out.njk`.