PP-7561 Render errors using response function
Don't call res.render directly for error pages, as we want to run them
through the display converter so we can display the logged in header
if the user is logged in.

For places where we did this, call `next` with an error and let the
`error-handler` render an appropriate error page using the function
in `response.js`.