PP-7701 Show notification on my services page
- Show a notification banner when users land on the My services page
to tell them about the behavioural change to have this as the default
page.

- When a "Hide message" button is clicked, set a cookie and hide the
banner.
- On subsequent page loads, check the cookie to decide whether to show
the notification or not.

- This works for both javascript enabled and not enabled. For javascript
enabled the banner is hidden and the cookie is set by client-side
javascript. If the javascript is not enabled the form inside the banner
is submitted to a route that sets the cookie then redirects to re-display
the my services page.

- Enable the `ENABLE_MY_SERVICES_AS_DEFAULT_VIEW` flag for Cypress tests
and update cypress tests for this behavioural change so we can test the
notification banner.

- Have not added environment variable check on date to show banner until
 yet.