PP-7701 Don't show notification after specified date
Check if the current date is after the date specified by the
`MY_SERVICES_DEFAULT_NOTIFICATION_END_DATE` environment variable, which
is expected to be a unix timestamp. If it is, do not display the
notification.

If the environment variable is not set, show the notification if the
other conditions are met.

Skipping adding tests for this as we expect the banner to be very short
lived.