PP-7708 List team members with take telephone payments roles
On the page for managing team members, list users who have the
view-and-initiate-moto or view-refund-and-initiate-moto roles.

It’s not currently possible to assign these roles to users so
there should not be any yet.

This contains a temporary change to the getRoleByExtId function
to hide the new roles from the code for managing team members and
inviting new users so someone cannot sneakily assign these new
roles to users before we’re ready.