PP-7708 Allow assigning existing users to telephone payments roles
Allow admin users of a service to change an existing user’s role
to view-and-initiate-moto or view-refund-and-initiate-moto.

For now, these two new roles are only shown if the service has the
agent_initiated_moto_enabled flag set, which we’re using to limit
which services have the agent-initiated MOTO payments feature.

If the flag is false, a determined admin could still assign one of
the new roles to a user. However, on its own that will not allow
the user to to do anything more, so it’s not really worth trying to
prevent this.