BAU Fix logged out error shown for phone number page
We now decide whether to show the logged in or logged out header in the
display converter, so that we can show the logged in header for errors
(PP-7561).

In this case, we were rendering the view directly using `res.render`
rather than using the `response` function to render after calling the
display converter.

This was the only case for a logged in route where we were rendering in
this way.