PP-7708 Allow inviting new users to telephone payments roles
Allow admin users of a service to invite a user granting them the
view-and-initiate-moto or view-refund-and-initiate-moto role.

For now, these two new roles are only shown if the service has the
agent_initiated_moto_enabled flag set, which we’re using to limit
which services have the agent-initiated MOTO payments feature.

If the flag is false, a determined admin could still invite a user
and grant them one of the new roles. However, on its own that will
not allow the user to to do anything more, so it’s not really worth
trying to prevent this.