PP-7786: update payouts page with status filter
Allow the payments to your bank account page to filter live or test
accounts based on the updated cross account permissions utility.

Add a tag to show what environment is being displayed following the
ticket design.

Update the download link for live and test environments.

Add a link to switch between live and test environments with the
following logic:
* viewing test data and has live accounts - show switch to live
* viewing test data with no live accounts - nothing
* viewing live data - show switch to test